This commit: * Adds SPJ_ prefix to ndbinfo.counter counters that are fetched from the SPJ block. * Attempts to remove warnings concerning implict conversions from size_t to Uint32   on Windows 7.

+++ storage/ndb/include/ndbapi/NdbQueryBuilder.hpp
@@ -306,7 +306,7 @@ public:
   // Fixed sized datatypes requires 'len' to exactly match the destination column.
   // Fixed sized character values should be spacepadded to required length.
   // Variable sized datatypes requires 'len <= max(len)'.
-  NdbConstOperand* constValue(const void* value, size_t len);
+  NdbConstOperand* constValue(const void* value, Uint32 len);
 
   // ::paramValue() is a placeholder for a parameter value to be specified when
   // a query instance is created for execution.
